Rename tokenstat module to license_report
