+++ agent/lm_agent/exceptions.py
@@ -28,8 +28,10 @@ class LicenseManagerNonSupportedServerTypeError(Buzz):
 class LicenseManagerBadServerOutput(Buzz):
     """Exception for license server bad output."""
 
+
 class CommandFailedToExecute(Buzz):
     """Exception for failed execution of command."""
+
 
 class LicenseManagerFeatureConfigurationIncorrect(Buzz):
     """Exception for feature configuration incorrectly formatted."""
