+++ agent/lm_agent/reconciliation.py
@@ -19,7 +19,7 @@ from lm_agent.exceptions import (
     LicenseManagerFeatureConfigurationIncorrect,
 )
 from lm_agent.logs import logger
-from lm_agent.tokenstat import report
+from lm_agent.license_report import report
 from lm_agent.workload_managers.slurm.cmd_utils import (
     get_all_product_features_from_cluster,
     get_cluster_name,
